bot_async_factory/app/requirements.txt
redis 
asyncio 
python-dotenv 
loguru